# expression: implement vectorized evaluation for builtinFromUnixTime1ArgSig